Move board routes to a single file
