+++ src/ej_boards/forms.py
@@ -8,11 +8,9 @@ class BoardForm(forms.ModelForm):
         model = models.Board
         fields = ['slug', 'title', 'description']
 
-    def save(self, commit=True):
+    def _save(self, commit=True):
         instance = super(BoardForm, self).save(commit=False)
         instance.slug = slugify(instance.slug)
-
         if commit:
             instance.save()
-
         return instance
